start of my new dungeon generation system
Started working a 3rd version of dungeon generator but a visual bug appeared where none of my sprites would be shown in the camera display. Ill try to find a way to fix it but if not I might have to find something else because so far I haven't found a direct solution.